Move search_index function to toshi-query
