+++ toshi-query/Cargo.toml
@@ -7,7 +7,9 @@ license = "MIT"
 edition = "2018"
 
 [dependencies]
-tantivy                = "^0.10"
-failure                = "^0.1"
-serde                  = { version = "^1.0", features = ["derive"] }
-serde_json             = "^1.0"
+tantivy       = "^0.10"
+failure       = "^0.1"
+serde         = { version = "^1.0", features = ["derive"] }
+serde_json    = "^1.0"
+log           = { version = "0.4", features = ["std"] }
+
